+++ src/app/modules/shared/components/cedar-embeddable-metadata-editor/cedar-embeddable-metadata-editor.component.spec.ts
@@ -75,9 +75,6 @@ describe('CedarEmbeddableMetadataEditorComponent config', () => {
     'expandedInstanceYaml',
     'expandedDataQualityReport',
     'expandedSampleTemplateLinks',
-    'collapseStaticComponents',
-    'showStaticText',
-    'showAllMultiInstanceValues',
     'showTemplateDescription',
     'readOnlyMode',
   ];