+++ src/app/modules/shared/models/component/null-component.model.ts
@@ -1,13 +1,11 @@
 import { CedarComponent } from './cedar-component.model';
 import { LabelInfo } from '../info/label-info.model';
-import { StaticFieldComponent } from '../static/static-field-component.model';
 
 export class NullComponent implements CedarComponent {
   className = 'NullComponent';
   labelInfo: LabelInfo = new LabelInfo();
   name = '';
   path: string[] = [];
-  linkedStaticFieldComponent: StaticFieldComponent | null = null;
   hidden = false;
   hiddenInTemplate = false;