+++ src/app/modules/shared/models/element/abstract-element-component.model.ts
@@ -1,7 +1,6 @@
 import { CedarComponent } from '../component/cedar-component.model';
 import { ElementComponent } from '../component/element-component.model';
 import { LabelInfo } from '../info/label-info.model';
-import { StaticFieldComponent } from '../static/static-field-component.model';
 
 export abstract class AbstractElementComponent implements ElementComponent {
   className = 'AbstractElementComponent';
@@ -9,7 +8,6 @@ export abstract class AbstractElementComponent implements ElementComponent {
   path: string[] = [];
   children: CedarComponent[] = [];
   labelInfo: LabelInfo = new LabelInfo();
-  linkedStaticFieldComponent: StaticFieldComponent | null = null;
   hidden = false;
 
   /**