+++ src/app/modules/shared/models/field/abstract-field-component.model.ts
@@ -4,7 +4,6 @@ import { ValueInfo } from '../info/value-info.model';
 import { NumberInfo } from '../info/number-info.model';
 import { ChoiceInfo } from '../info/choice-info.model';
 import { LabelInfo } from '../info/label-info.model';
-import { StaticFieldComponent } from '../static/static-field-component.model';
 import { ControlledInfo } from '../info/controlled-info.model';
 import { MultiInfo } from '../info/multi-info.model';
 
@@ -17,7 +16,6 @@ export abstract class AbstractFieldComponent implements FieldComponent {
   numberInfo: NumberInfo = new NumberInfo();
   choiceInfo: ChoiceInfo = new ChoiceInfo();
   labelInfo: LabelInfo = new LabelInfo();
-  linkedStaticFieldComponent: StaticFieldComponent | null = null;
   controlledInfo: ControlledInfo = new ControlledInfo();
   multiInfo: MultiInfo = new MultiInfo();
   hidden = false;